requirements.txt
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
distlib==0.3.9
filelock==3.18.0
fsspec==2025.5.1
gdown==5.2.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gTTS==2.5.4
httplib2==0.22.0
identify==2.6.12
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.2
openai-whisper==20240930
pillow==11.2.1
pip-system-certs==4.0
platformdirs==4.3.8
praw==7.8.1
prawcore==2.4.0
pre_commit==4.2.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
soupsieve==2.7
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250516
typing_extensions==4.13.2
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.4.0
virtualenv==20.31.2
websocket-client==1.8.0
wrapt==1.17.2
